fix(header): names, style header, and new svgs
