+++ src/components/Layout.jsx
@@ -1,5 +1,5 @@
 // src/components/Layout.jsx
-import Navbar from "./Navbar";
+import Header from "./Header";
 import Footer from "./Footer";
 import styles from "./styles/Layout.module.css";
 import { Outlet } from "react-router-dom";
@@ -10,7 +10,7 @@ export default function Layout() {
   return (
     <div className={styles.wrapper}>
       <ScrollToTop />
-      <Navbar />
+      <Header />
       <main className={styles.main}>
         <Outlet />
       </main>
