+++ src/components/styles/Header.module.css
@@ -1,0 +1,186 @@
+.header {
+  background: rgba(0, 0, 0, 0.95);
+  backdrop-filter: blur(10px);
+  border-bottom: 1px solid #374151;
+  position: sticky;
+  top: 0;
+  z-index: 100;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  height: 4rem;
+  padding: 0 1rem;
+}
+
+.logo {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+}
+
+.logoIcon {
+  padding: 0.25rem 0.5rem;
+  border-radius: 0.5rem;
+  color: #000;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  max-width: 10rem;
+  max-height: 3.5rem;
+  overflow: hidden;
+}
+
+.logoIcon img {
+  display: block;
+  width: 100%;
+  height: 100%;
+  object-fit: contain;
+}
+
+.logoTitle {
+  font-size: 1.25rem;
+  font-weight: bold;
+  background: linear-gradient(135deg, #f59e0b, #f97316);
+  -webkit-background-clip: text;
+  background-clip: text;
+  color: transparent;
+  margin: 0;
+  white-space: nowrap;
+}
+
+.logoSubtitle {
+  font-size: 0.75rem;
+  color: #9ca3af;
+  margin: 0;
+  white-space: nowrap;
+}
+
+.navbar {
+  display: none; /* Por defecto oculto en móvil */
+  gap: 2rem;
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.navbar li a {
+  color: #fff;
+  text-decoration: none;
+  transition: color 0.2s;
+}
+
+.navbar li a:hover {
+  color: #f59e0b;
+}
+
+.headerActions {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.btnGhost {
+  background: transparent;
+  border: none;
+  cursor: pointer;
+  padding: 0.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.btnGhost img {
+  height: 1.5rem;
+  width: auto;
+}
+
+.btnGhostCartBtn {
+  position: relative;
+  background: transparent;
+  border: none;
+  cursor: pointer;
+  padding: 0.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.imgCart {
+  height: 1.5rem;
+  width: auto;
+}
+
+.imgSearch {
+  height: 1.5rem;
+  width: auto;
+}
+
+.imgMenu {
+  height: 1.5rem;
+  width: auto;
+}
+
+.cartBadge {
+  position: absolute;
+  top: -0.25rem;
+  right: -0.25rem;
+  background: #f59e0b;
+  color: #000;
+  font-size: 0.75rem;
+  font-weight: bold;
+  padding: 0.125rem 0.375rem;
+  border-radius: 9999px;
+  min-width: 1.25rem;
+  text-align: center;
+}
+
+.btnGhostMenuBtn {
+  display: block;
+  background: transparent;
+  border: none;
+  cursor: pointer;
+  padding: 0.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.scrolled {
+  background: rgba(0, 0, 0, 0.7);
+  backdrop-filter: blur(20px);
+}
+
+@media (min-width: 768px) {
+  .navbar {
+    display: flex;
+    /* Elimina los margin-auto si no te dan el centrado visual deseado */
+    /* margin-left: auto; */
+    /* margin-right: auto; */
+
+    /* Centra el navbar directamente en el espacio restante */
+    flex-grow: 1; /* Permite que ocupe el espacio disponible */
+    justify-content: center; /* Centra sus propios elementos */
+
+    /* Desplaza el navbar a la izquierda para compensar el logo */
+    transform: translateX(
+      -50px
+    ); /* Ajusta este valor (ej: -50px, -75px, -100px) */
+  }
+  .btnGhostMenuBtn {
+    display: none;
+  }
+}
+
+@media (max-width: 767px) {
+  .navOpen {
+    display: flex;
+    position: absolute;
+    top: 100%;
+    left: 0;
+    right: 0;
+    background: rgba(0, 0, 0, 0.95);
+    flex-direction: column;
+    padding: 1rem;
+    border-bottom: 1px solid #374151;
+  }
+}
