--- src/components/styles/Navbar.module.css
@@ -1,158 +1,0 @@
-/* Envoltura completa: incluye topbar + navbar */
-.navbarWrapper {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  z-index: 1000;
-}
-
-/* Barra naranja superior */
-.navTop {
-  background-color: #f97316; /* naranja */
-  color: white;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.3rem 1rem;
-  font-size: 0.85rem;
-}
-
-.redes a {
-  margin-right: 0.5rem;
-}
-
-.redes img {
-  width: 18px;
-  height: 18px;
-  opacity: 0.9;
-  transition: transform 0.2s ease;
-}
-
-.redes img:hover {
-  transform: scale(1.1);
-  opacity: 1;
-}
-
-.telefono {
-  font-weight: 500;
-}
-
-/* Navbar principal */
-.navbar {
-  background-color: rgba(30, 41, 59, 0.95);
-  backdrop-filter: blur(6px);
-  border-bottom: 1px solid #334155;
-  transition: background-color 0.3s ease, backdrop-filter 0.3s ease;
-}
-
-/* Efecto al hacer scroll */
-.scrolled .navbar {
-  background-color: rgba(30, 41, 59, 0.8);
-  backdrop-filter: blur(10px);
-}
-
-.navInner {
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 0.75rem 1.5rem;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  flex-wrap: wrap;
-}
-
-/* Logo a la izquierda */
-.logo {
-  height: 2.5rem;
-  width: auto;
-}
-
-/* Links centrados */
-.navLinks {
-  list-style: none;
-  display: flex;
-  gap: 1rem;
-  justify-content: center;
-  flex: 1;
-}
-
-.navLinks li {
-  display: flex;
-}
-
-.navLinks a {
-  text-decoration: none;
-  color: #e2e8f0;
-  font-weight: 500;
-  padding: 0.4rem 1rem;
-  border-radius: 6px;
-  background-color: #334155;
-  transition: all 0.2s ease;
-}
-
-.navLinks a:hover {
-  background-color: #475569;
-  color: #facc15;
-}
-
-/* Iconos a la derecha */
-.navIcons {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.imgCart,
-.imgUser {
-  width: 24px;
-  height: 24px;
-  cursor: pointer;
-  opacity: 0.9;
-  transition: transform 0.2s ease;
-}
-
-.imgCart:hover,
-.imgUser:hover {
-  transform: scale(1.1);
-  opacity: 1;
-}
-
-/* Responsive: stack navLinks y shrink layout */
-@media (max-width: 768px) {
-  .navInner {
-    flex-direction: column;
-    align-items: stretch;
-  }
-
-  .navLinks {
-    flex-direction: column;
-    align-items: center;
-    padding-top: 0.5rem;
-  }
-
-  .navIcons {
-    justify-content: center;
-    margin-top: 0.5rem;
-  }
-
-  .logo {
-    margin-bottom: 0.5rem;
-  }
-
-  .navTop {
-    flex-direction: column;
-    gap: 0.25rem;
-    text-align: center;
-  }
-
-  .telefono {
-    margin-top: 0.3rem;
-  }
-}
-
-.telefono a {
-  color: white;
-  text-decoration: none;
-  font-weight: bold;
-}
